Refactor Account related routes
