+++ packages/bierzo-wallet/src/routes/account/index.ts
@@ -1,0 +1,5 @@
+import AccountManage from "./manage";
+import AccountRegister from "./register";
+import AccountUpdate from "./update";
+
+export { AccountManage, AccountRegister, AccountUpdate };
