+++ packages/bierzo-wallet/src/routes/account/register/components/ConfirmRegistration.tsx
