+++ packages/bierzo-wallet/src/routes/account/register/components/NameForm.tsx
@@ -1,4 +1,5 @@
 import { Fee, Identity, TransactionId } from "@iov/bcp";
+import { bnsCodec } from "@iov/bns";
 import { JsonRpcRequest } from "@iov/jsonrpc";
 import { FieldValidator } from "final-form";
 import {
@@ -22,31 +23,31 @@ import {
 import React from "react";
 
 import {
-  generateRegisterUsernameTxRequest,
-  generateUpdateUsernameTxRequest,
-} from "../../../communication/requestgenerators";
-import { RpcEndpoint } from "../../../communication/rpcEndpoint";
+  generateRegisterAccountTxRequest,
+  generateReplaceAccountTargetsTxRequest,
+} from "../../../../communication/requestgenerators";
+import { RpcEndpoint } from "../../../../communication/rpcEndpoint";
 import {
   getAddressItems,
   getChainAddressPairsFromValues,
   getFormInitValues,
   getSubmitButtonCaption,
-} from "../../../components/AccountEdit";
+} from "../../../../components/AccountEdit";
 import {
   AddressesTooltipHeader,
   BwUsernameWithChainName,
   TooltipContent,
-} from "../../../components/AccountManage";
-import { AddressesTableProps } from "../../../components/AddressesTable";
-import LedgerBillboardMessage from "../../../components/BillboardMessage/LedgerBillboardMessage";
-import NeumaBillboardMessage from "../../../components/BillboardMessage/NeumaBillboardMessage";
-import PageContent from "../../../components/PageContent";
-import { isValidIov } from "../../../logic/account";
+} from "../../../../components/AccountManage";
+import { AddressesTableProps } from "../../../../components/AddressesTable";
+import LedgerBillboardMessage from "../../../../components/BillboardMessage/LedgerBillboardMessage";
+import NeumaBillboardMessage from "../../../../components/BillboardMessage/NeumaBillboardMessage";
+import PageContent from "../../../../components/PageContent";
+import { isValidName } from "../../../../logic/account";
 import shield from "../assets/shield.svg";
 import SelectAddressesTable from "./SelectAddressesTable";
 
-export const REGISTER_IOVNAME_VIEW_ID = "register-iovname-view-id";
-export const REGISTER_IOVNAME_FIELD = "register-iovname-field";
+export const REGISTER_NAME_VIEW_ID = "register-name-view-id";
+export const REGISTER_NAME_FIELD = "register-name-field";
 
 const registerIcon = <Image src={shield} alt="shield" />;
 
@@ -56,12 +57,12 @@ const useStyles = makeStyles({
   },
 });
 
-export function NoIovnameHeader(): JSX.Element {
+function NoNameHeader(): JSX.Element {
   const classes = useStyles();
   return (
     <Block className={classes.iovnameHeader} borderRadius={40} width={145} padding={1}>
       <Typography variant="subtitle1" weight="semibold" color="primary" align="center">
-        yourname*iov
+        eg. anna*yourstarname
       </Typography>
     </Block>
   );
@@ -70,13 +71,13 @@ export function NoIovnameHeader(): JSX.Element {
 interface Props extends AddressesTableProps {
   readonly onCancel: () => void;
   readonly iovnameAddresses: BwUsernameWithChainName | undefined;
-  readonly bnsIdentity: Identity | undefined;
-  readonly rpcEndpoint: RpcEndpoint | undefined;
+  readonly bnsIdentity: Identity;
+  readonly rpcEndpoint: RpcEndpoint;
   readonly transactionFee: Fee | undefined;
   readonly setTransactionId: React.Dispatch<React.SetStateAction<TransactionId | null>>;
 }
 
-const IovnameForm = ({
+const NameForm = ({
   chainAddresses,
   iovnameAddresses,
   bnsIdentity,
@@ -96,25 +97,26 @@ const IovnameForm = ({
   }, [chainAddresses, iovnameAddresses]);
 
   const onSubmit = async (values: object): Promise<void> => {
-    if (!bnsIdentity) throw Error("No bnsIdentity found for submit");
-    if (!rpcEndpoint) throw Error("No rpcEndpoint found for submit");
-
     const formValues = values as FormValues;
 
     const addressesToRegister = getChainAddressPairsFromValues(formValues, chainAddresses);
+    const [name, domain] = formValues[REGISTER_NAME_FIELD].split("*");
 
     try {
       let request: JsonRpcRequest;
       if (iovnameAddresses) {
-        request = await generateUpdateUsernameTxRequest(
+        request = await generateReplaceAccountTargetsTxRequest(
           bnsIdentity,
-          iovnameAddresses.username,
+          domain,
+          name,
           addressesToRegister,
         );
       } else {
-        request = await generateRegisterUsernameTxRequest(
+        request = await generateRegisterAccountTxRequest(
           bnsIdentity,
-          formValues[REGISTER_IOVNAME_FIELD],
+          domain,
+          name,
+          bnsCodec.identityToAddress(bnsIdentity),
           addressesToRegister,
         );
       }
@@ -149,61 +151,58 @@ const IovnameForm = ({
     }
   };
 
-  const iovnameValidator: FieldValidator<FieldInputValue> = (value): string | undefined => {
+  const nameValidator: FieldValidator<FieldInputValue> = (value): string | undefined => {
     if (!iovnameAddresses) {
       if (!value) {
         return "Required";
       }
 
-      const checkResult = isValidIov(value);
+      const checkResult = isValidName(value);
 
       switch (checkResult) {
-        case "not_iov":
-          return "Iovname must end with *iov";
         case "wrong_number_of_asterisks":
-          return "Iovname must include only one namespace";
+          return "Name must include only one namespace";
         case "too_short":
-          return "Iovname should be at least 3 characters";
+          return "Name should be at least 3 characters";
         case "too_long":
-          return "Iovname should be maximum 64 characters";
+          return "Name should be maximum 64 characters";
         case "wrong_chars":
-          return "Iovname should contain 'abcdefghijklmnopqrstuvwxyz0123456789-_.' characters only";
+          return "Name should contain 'abcdefghijklmnopqrstuvwxyz0123456789-_.' characters only";
         case "valid":
           break;
         default:
-          throw new Error(`"Unknown iovname validation error: ${checkResult}`);
+          throw new Error(`"Unknown name validation error: ${checkResult}`);
       }
     }
 
     return undefined;
   };
 
-  /* const validateIovnameExistsAndAddresses = React.useCallback(
+  /* const validateNameExistsAndAddresses = React.useCallback(
     (values: object) => {
       const validate = async (values: object): Promise<object> => {
         const formValues = values as FormValues;
         const errors: ValidationError = {};
 
         const connection = await getConnectionForBns();
-        const username = formValues[REGISTER_IOVNAME_FIELD];
-        const usernames = await connection.getUsernames({ username: username });
-        if (usernames.length > 0) {
-          errors[REGISTER_IOVNAME_FIELD] = "Iovname already exists";
+        const [name, domain] = formValues[REGISTER_NAME_FIELD].split("*");
+        const accounts = await connection.getAccounts({ domain: domain });
+        if (accounts.find(account => account.name === name)) {
+          errors[REGISTER_NAME_FIELD] = "Name already exists";
           return errors;
         }
 
         const addressesToRegister = getChainAddressPairsFromValues(formValues, chainAddresses);
         for (const address of addressesToRegister) {
-          try {
-            const codec = await getCodecForChainId(address.chainId);
-            if (!codec.isValidAddress(address.address)) {
-              const addressField = Object.entries(formValues).find(([_id, value]) => value === address.address);
-              if (addressField) {
-                errors[addressField[0]] = "Not valid blockchain address";
-              }
+          const codec = await getCodecForChainId(address.chainId);
+          if (!codec.isValidAddress(address.address)) {
+            const addressField = Object.entries(formValues).find(([_id, value]) => {
+              if (value === address.address) return true;
+              return false;
+            });
+            if (addressField) {
+              errors[addressField[0]] = "Not valid blockchain address";
             }
-          } catch (err) {
-            console.info(err);
           }
         }
 
@@ -218,7 +217,7 @@ const IovnameForm = ({
   const initialValues = React.useMemo(() => getFormInitValues(chainAddressesItems), [chainAddressesItems]);
   const { form, handleSubmit, invalid, submitting, validating } = useForm({
     onSubmit,
-    // validate: validateIovnameExistsAndAddresses,
+    // validate: validateNameExistsAndAddresses,
     initialValues,
   });
 
@@ -251,7 +250,7 @@ const IovnameForm = ({
 
   return (
     <Form onSubmit={handleSubmit}>
-      <PageContent id={REGISTER_IOVNAME_VIEW_ID} icon={registerIcon} buttons={buttons} avatarColor="#31E6C9">
+      <PageContent id={REGISTER_NAME_VIEW_ID} icon={registerIcon} buttons={buttons} avatarColor="#31E6C9">
         <Block textAlign="left">
           {iovnameAddresses && (
             <Typography variant="h4" align="center">
@@ -262,11 +261,11 @@ const IovnameForm = ({
             <React.Fragment>
               <Block display="flex" justifyContent="space-between" marginBottom={1}>
                 <Typography variant="subtitle2" weight="semibold">
-                  Create your iovname
+                  Register your new name
                 </Typography>
                 <Block display="flex" alignItems="center">
                   <Tooltip maxWidth={320}>
-                    <TooltipContent header={<NoIovnameHeader />} title="Choose your address">
+                    <TooltipContent header={<NoNameHeader />} title="Choose your address">
                       With IOV you can choose your easy to read human readable address. No more complicated
                       cryptography when sending to friends.
                     </TooltipContent>
@@ -279,9 +278,9 @@ const IovnameForm = ({
               </Block>
               <Block width="100%" marginBottom={1}>
                 <TextField
-                  name={REGISTER_IOVNAME_FIELD}
+                  name={REGISTER_NAME_FIELD}
                   form={form}
-                  validate={iovnameValidator}
+                  validate={nameValidator}
                   placeholder="eg. yourname*iov"
                   fullWidth
                   margin="none"
@@ -302,7 +301,7 @@ const IovnameForm = ({
                 <Tooltip maxWidth={320}>
                   <TooltipContent header={<AddressesTooltipHeader />} title="Your linked addresses">
                     With Neuma you can have an universal blockchain address that is linked to all your
-                    addresses. Just give your friends your iovname.
+                    addresses. Just give your friends your starname.
                   </TooltipContent>
                 </Tooltip>
               </Block>
@@ -325,4 +324,4 @@ const IovnameForm = ({
   );
 };
 
-export default IovnameForm;
+export default NameForm;
